fix: 解决 Kotlin 编译错误（共9处）
根据 GitHub Actions 错误日志逐一修复:
1. 🔧 WallpaperLayer: 添加 ALBUM_ART 分支（when exhaustive）
2. 🔧 WaterRipple: 添加 import androidx.compose.ui.unit.dp
3. 🔧 FlipClock: 解决 TextStyle 导入冲突（compose vs java.time）
4. 🔧 FlipClock: 修复 TextStyle.FULL → JTextStyle.FULL
5. 🔧 AppNavigation: 修复 .align() → .align(Alignment.BottomCenter)
6. 🔧 ClockScreen: 添加 import Alignment
7. 🔧 AppTheme: 添加 import ThemeSettings
8. 🔧 MediaPlaybackMonitor: 修复 Int/Float 类型不匹配
9. 🔧 所有修复参照官方 Gradle/Android 文档

+++ app/src/main/java/com/zhuomo/glasspomodoro/media/MediaPlaybackMonitor.kt
@@ -228,7 +228,7 @@ class AlbumArtColorExtractor {
                 val pg = android.graphics.Color.green(pixel)
                 val pb = android.graphics.Color.blue(pixel)
 
-                var minDist = Float.MAX_VALUE
+                var minDist = Int.MAX_VALUE
                 var bestIdx = 0
                 clusters.forEachIndexed { idx, cluster ->
                     val dr = pr - (if (cluster.count > 0) cluster.r / cluster.count else cluster.r).coerceIn(0, 255)
